+++ lib/authentication/log_in.dart
@@ -1,14 +1,15 @@
-// ignore_for_file: use_build_context_synchronously
+// ignore_for_file: use_build_context_synchronously, use_key_in_widget_constructors, library_private_types_in_public_api, unused_field, prefer_typing_uninitialized_variables, no_leading_underscores_for_local_identifiers
 
 import 'dart:convert';
-import 'package:shared_preferences/shared_preferences.dart';
 import 'package:attendifyyy/api_connection/api_connection.dart';
 import 'package:attendifyyy/authentication/sign_up.dart';
 import 'package:attendifyyy/authentication/user_preferences/user_preferences.dart';
 import 'package:attendifyyy/bottom_nav_bar.dart';
+import 'package:attendifyyy/utils/common_widgets/text_fields.dart';
 import 'package:flutter/material.dart';
 import 'package:http/http.dart' as http;
 import 'package:email_validator/email_validator.dart';
+import 'package:shared_preferences/shared_preferences.dart';
 
 class LogIn extends StatefulWidget {
   @override
@@ -39,17 +40,17 @@ class _LogInState extends State<LogIn> {
 
   //in first page load if there is an existing credentials stored in local storage
   //occupy the email and password controller with those values
- Future<void> fillInLoginCredentials() async {
+  Future<void> fillInLoginCredentials() async {
     final SharedPreferences _prefs = await SharedPreferences.getInstance();
 
-      if(_prefs.getString('remember_email') != null &&
-          _prefs.getString('remember_pass') != null) {
-        setState(() {
+    if (_prefs.getString('remember_email') != null &&
+        _prefs.getString('remember_pass') != null) {
+      setState(() {
         emailController.text = _prefs.getString('remember_email')!;
         passwordController.text = _prefs.getString('remember_pass')!;
         rememberMe = true;
-        });
-      }
+      });
+    }
   }
 
   //post credential to database
@@ -71,13 +72,13 @@ class _LogInState extends State<LogIn> {
           MaterialPageRoute(builder: (context) => const BottomNavBar()),
         );
         ScaffoldMessenger.of(context)
-            .showSnackBar(SnackBar(content: Text("Log in successfully")));
+            .showSnackBar(const SnackBar(content: Text("Log in successfully")));
 
         await RememberUserPreferences.storeUserInfo(
             resBodyOfLogin["teacherData"]);
       } else {
         ScaffoldMessenger.of(context)
-            .showSnackBar(SnackBar(content: Text("Invalid Credentials")));
+            .showSnackBar(const SnackBar(content: Text("Invalid Credentials")));
       }
     } else {
       setState(() {
@@ -134,16 +135,20 @@ class _LogInState extends State<LogIn> {
                 width: 200,
                 fit: BoxFit.contain,
               ),
-              createTextField(emailController, 'Email', (value) {
-                if (value == null || value.isEmpty) {
-                  return "This field is required.";
-                } else if (!EmailValidator.validate(value)) {
-                  return "Please use a valid email address.";
-                }
-                return null;
-              }),
+              createTextField(
+                emailController,
+                'Email',
+                (value) {
+                  if (value == null || value.isEmpty) {
+                    return "This field is required.";
+                  } else if (!EmailValidator.validate(value)) {
+                    return "Please use a valid email address.";
+                  }
+                  return null;
+                },
+              ),
               const SizedBox(height: 20),
-              //Password with obscure sht
+              // lahi nga textformfield paras password with obscure sht
               TextFormField(
                 controller: passwordController,
                 obscureText: obscurePassword,
@@ -153,24 +158,44 @@ class _LogInState extends State<LogIn> {
                   labelStyle: const TextStyle(
                       color: Color(0xFFABABAB),
                       fontSize: 14), //affect the size of textfield
-                  floatingLabelStyle: const TextStyle(color: Color(0xFF081631)),
+                  floatingLabelStyle: const TextStyle(
+                    color: Color(0xFF081631),
+                  ),
                   //when textField is focused or selected
                   focusedBorder: const OutlineInputBorder(
-                      borderRadius: BorderRadius.all(Radius.circular(10.0)),
-                      borderSide:
-                          BorderSide(width: 2, color: Color(0xFF081631))),
-                  //normal state of textField border
+                    borderRadius: BorderRadius.all(
+                      Radius.circular(10.0),
+                    ),
+                    borderSide: BorderSide(
+                      width: 2,
+                      color: Color(0xFF081631),
+                    ),
+                  ),
                   enabledBorder: const OutlineInputBorder(
-                      borderRadius: BorderRadius.all(Radius.circular(10.0)),
-                      borderSide:
-                          BorderSide(color: Color(0xFFABABAB))), // your color
+                    borderRadius: BorderRadius.all(
+                      Radius.circular(10.0),
+                    ),
+                    borderSide: BorderSide(
+                      color: Color(0xFFABABAB),
+                    ),
+                  ),
                   errorBorder: const OutlineInputBorder(
-                      borderRadius: BorderRadius.all(Radius.circular(10.0)),
-                      borderSide: BorderSide(color: Color(0xFFFF0000))),
+                    borderRadius: BorderRadius.all(
+                      Radius.circular(10.0),
+                    ),
+                    borderSide: BorderSide(
+                      color: Color(0xFFFF0000),
+                    ),
+                  ),
                   focusedErrorBorder: const OutlineInputBorder(
-                      borderRadius: BorderRadius.all(Radius.circular(10.0)),
-                      borderSide:
-                          BorderSide(width: 2, color: Color(0xFFFF0000))),
+                    borderRadius: BorderRadius.all(
+                      Radius.circular(10.0),
+                    ),
+                    borderSide: BorderSide(
+                      width: 2,
+                      color: Color(0xFFFF0000),
+                    ),
+                  ),
                   suffixIcon: GestureDetector(
                     onTap: () {
                       setState(() {
@@ -194,18 +219,23 @@ class _LogInState extends State<LogIn> {
               Row(
                 children: [
                   Checkbox(
-                      value: rememberMe,
-                      activeColor: const Color(0xFF081631),
-                      onChanged: (value) {
-                        setState(() {
+                    value: rememberMe,
+                    activeColor: const Color(0xFF081631),
+                    onChanged: (value) {
+                      setState(
+                        () {
                           rememberMe = !rememberMe;
-                        });
-                      }),
-                  const Text("Remember Me",
-                      style: TextStyle(
-                          color: Color(0xFF081631),
-                          fontSize: 13.0,
-                          fontWeight: FontWeight.bold))
+                        },
+                      );
+                    },
+                  ),
+                  const Text(
+                    "Remember Me",
+                    style: TextStyle(
+                        color: Color(0xFF081631),
+                        fontSize: 13.0,
+                        fontWeight: FontWeight.bold),
+                  )
                 ],
               ),
               const SizedBox(height: 40),
@@ -216,20 +246,26 @@ class _LogInState extends State<LogIn> {
                   }
                 },
                 style: ButtonStyle(
-                    minimumSize: MaterialStateProperty.all<Size>(
-                        const Size.fromHeight(60)),
-                    backgroundColor: MaterialStateProperty.all<Color>(
-                        const Color(0xFF081631)),
-                    shape: MaterialStateProperty.all<RoundedRectangleBorder>(
-                        RoundedRectangleBorder(
+                  minimumSize: MaterialStateProperty.all<Size>(
+                    const Size.fromHeight(60),
+                  ),
+                  backgroundColor: MaterialStateProperty.all<Color>(
+                    const Color(0xFF081631),
+                  ),
+                  shape: MaterialStateProperty.all<RoundedRectangleBorder>(
+                    RoundedRectangleBorder(
                       borderRadius: BorderRadius.circular(10.0),
-                    ))),
-                child: const Text('LOG IN',
-                    style: TextStyle(
-                      fontSize: 16.0,
-                      color: Colors.white,
-                      fontWeight: FontWeight.bold,
-                    )),
+                    ),
+                  ),
+                ),
+                child: const Text(
+                  'LOG IN',
+                  style: TextStyle(
+                    fontSize: 16.0,
+                    color: Colors.white,
+                    fontWeight: FontWeight.bold,
+                  ),
+                ),
               ),
               const SizedBox(height: 8),
               Row(
@@ -240,20 +276,26 @@ class _LogInState extends State<LogIn> {
                     style: TextStyle(fontSize: 16.0, color: Color(0xFF777777)),
                   ),
                   TextButton(
-                      onPressed: () {
-                        Navigator.push(
-                          context,
-                          MaterialPageRoute(builder: (context) => SignUp()),
-                        );
-                      },
-                      style: ButtonStyle(
-                        //this padding is the distance between ...account? and Sign up text
-                        padding: MaterialStateProperty.all<EdgeInsets>(
-                            const EdgeInsets.all(5.0)),
+                    onPressed: () {
+                      Navigator.push(
+                        context,
+                        MaterialPageRoute(builder: (context) => SignUp()),
+                      );
+                    },
+                    style: ButtonStyle(
+                      //this padding is the distance between ...account? and Sign up text
+                      padding: MaterialStateProperty.all<EdgeInsets>(
+                        const EdgeInsets.all(5.0),
                       ),
-                      child: const Text('Sign up',
-                          style: TextStyle(
-                              fontSize: 16.0, color: Color(0xFF081631))))
+                    ),
+                    child: const Text(
+                      'Sign up',
+                      style: TextStyle(
+                        fontSize: 16.0,
+                        color: Color(0xFF081631),
+                      ),
+                    ),
+                  ),
                 ],
               )
             ],
@@ -262,39 +304,4 @@ class _LogInState extends State<LogIn> {
       ),
     );
   }
-}
-
-/*
-*
-* Widget components
-*
-*
-* */
-Widget createTextField(valueController, label, validationFunction) {
-  return TextFormField(
-    validator: validationFunction,
-    controller: valueController,
-    decoration: InputDecoration(
-      labelText: label,
-      contentPadding: const EdgeInsets.symmetric(horizontal: 15),
-      labelStyle: const TextStyle(
-          color: Color(0xFFABABAB),
-          fontSize: 14), //affect the size of textfield
-      floatingLabelStyle: const TextStyle(color: Color(0xFF081631)),
-      //when textField is focused or selected
-      focusedBorder: const OutlineInputBorder(
-          borderRadius: BorderRadius.all(Radius.circular(10.0)),
-          borderSide: BorderSide(width: 2, color: Color(0xFF081631))),
-      //normal state of textField border
-      enabledBorder: const OutlineInputBorder(
-          borderRadius: BorderRadius.all(Radius.circular(10.0)),
-          borderSide: BorderSide(color: Color(0xFFABABAB))), // your color
-      errorBorder: const OutlineInputBorder(
-          borderRadius: BorderRadius.all(Radius.circular(10.0)),
-          borderSide: BorderSide(color: Color(0xFFFF0000))),
-      focusedErrorBorder: const OutlineInputBorder(
-          borderRadius: BorderRadius.all(Radius.circular(10.0)),
-          borderSide: BorderSide(width: 2, color: Color(0xFFFF0000))),
-    ),
-  );
 }
