+++ lib/authentication/sign_up.dart
@@ -1,5 +1,8 @@
+// ignore_for_file: use_key_in_widget_constructors, library_private_types_in_public_api, use_build_context_synchronously
+
 import 'package:attendifyyy/api_connection/api_connection.dart';
 import 'package:attendifyyy/authentication/log_in.dart';
+import 'package:attendifyyy/utils/common_widgets/text_fields.dart';
 import 'package:flutter/material.dart';
 import 'package:http/http.dart' as http;
 import 'package:email_validator/email_validator.dart';
@@ -38,12 +41,17 @@ class _SignUpState extends State<SignUp> {
 
     if (response.statusCode == 200) {
       _response = response.body;
-      ScaffoldMessenger.of(context)
-          .showSnackBar(SnackBar(content: Text(_response)));
+      ScaffoldMessenger.of(context).showSnackBar(
+        SnackBar(
+          content: Text(_response),
+        ),
+      );
     } else {
-      setState(() {
-        _response = 'Error';
-      });
+      setState(
+        () {
+          _response = 'Error';
+        },
+      );
     }
   }
 
@@ -66,32 +74,44 @@ class _SignUpState extends State<SignUp> {
                   height: 200,
                   fit: BoxFit.contain,
                 ),
-                createTextField(firstNameController, 'First Name', (value) {
-                  if (value == null || value.isEmpty) {
-                    return "This field is required.";
-                  } else if (value.contains(RegExp(r'[0-9]'))) {
-                    return "Must contain only letters.";
-                  }
-                  return null;
-                }),
+                createTextField(
+                  firstNameController,
+                  'First Name',
+                  (value) {
+                    if (value == null || value.isEmpty) {
+                      return "This field is required.";
+                    } else if (value.contains(RegExp(r'[0-9]'))) {
+                      return "Must contain only letters.";
+                    }
+                    return null;
+                  },
+                ),
                 const SizedBox(height: 12),
-                createTextField(lastNameController, 'Last Name', (value) {
-                  if (value == null || value.isEmpty) {
-                    return "This field is required.";
-                  } else if (value.contains(RegExp(r'[0-9]'))) {
-                    return "Must contain only letters.";
-                  }
-                  return null;
-                }),
+                createTextField(
+                  lastNameController,
+                  'Last Name',
+                  (value) {
+                    if (value == null || value.isEmpty) {
+                      return "This field is required.";
+                    } else if (value.contains(RegExp(r'[0-9]'))) {
+                      return "Must contain only letters.";
+                    }
+                    return null;
+                  },
+                ),
                 const SizedBox(height: 12),
-                createTextField(emailController, 'Email', (value) {
-                  if (value == null || value.isEmpty) {
-                    return "This field is required.";
-                  } else if (!EmailValidator.validate(value)) {
-                    return "Please use a valid email address.";
-                  }
-                  return null;
-                }),
+                createTextField(
+                  emailController,
+                  'Email',
+                  (value) {
+                    if (value == null || value.isEmpty) {
+                      return "This field is required.";
+                    } else if (!EmailValidator.validate(value)) {
+                      return "Please use a valid email address.";
+                    }
+                    return null;
+                  },
+                ),
                 const SizedBox(height: 12),
                 //Password with obscure sht
                 TextFormField(
@@ -107,26 +127,31 @@ class _SignUpState extends State<SignUp> {
                         const TextStyle(color: Color(0xFF081631)),
                     //when textField is focused or selected
                     focusedBorder: const OutlineInputBorder(
-                        borderRadius: BorderRadius.all(Radius.circular(10.0)),
-                        borderSide:
-                            BorderSide(width: 2, color: Color(0xFF081631))),
+                      borderRadius: BorderRadius.all(Radius.circular(10.0)),
+                      borderSide:
+                          BorderSide(width: 2, color: Color(0xFF081631)),
+                    ),
                     //normal state of textField border
                     enabledBorder: const OutlineInputBorder(
-                        borderRadius: BorderRadius.all(Radius.circular(10.0)),
-                        borderSide:
-                            BorderSide(color: Color(0xFFABABAB))), // your color
+                      borderRadius: BorderRadius.all(Radius.circular(10.0)),
+                      borderSide: BorderSide(color: Color(0xFFABABAB)),
+                    ), // your color
                     errorBorder: const OutlineInputBorder(
-                        borderRadius: BorderRadius.all(Radius.circular(10.0)),
-                        borderSide: BorderSide(color: Color(0xFFFF0000))),
+                      borderRadius: BorderRadius.all(Radius.circular(10.0)),
+                      borderSide: BorderSide(color: Color(0xFFFF0000)),
+                    ),
                     focusedErrorBorder: const OutlineInputBorder(
-                        borderRadius: BorderRadius.all(Radius.circular(10.0)),
-                        borderSide:
-                            BorderSide(width: 2, color: Color(0xFFFF0000))),
+                      borderRadius: BorderRadius.all(Radius.circular(10.0)),
+                      borderSide:
+                          BorderSide(width: 2, color: Color(0xFFFF0000)),
+                    ),
                     suffixIcon: GestureDetector(
                       onTap: () {
-                        setState(() {
-                          obscurePassword = !obscurePassword;
-                        });
+                        setState(
+                          () {
+                            obscurePassword = !obscurePassword;
+                          },
+                        );
                       },
                       child: Icon(
                         obscurePassword
@@ -150,25 +175,33 @@ class _SignUpState extends State<SignUp> {
                   },
                 ),
                 const SizedBox(height: 12),
-                createTextField(phoneNumberController, 'Phone Number', (value) {
-                  if (value == null || value.isEmpty) {
-                    return "This field is required.";
-                  } else if (value.length <= 10) {
-                    return "Please input the correct number.";
-                  } else if (value.contains(RegExp(r'[A-Z, a-z]'))) {
-                    return "Must contain only numbers.";
-                  }
-                  return null;
-                }),
+                createTextField(
+                  phoneNumberController,
+                  'Phone Number',
+                  (value) {
+                    if (value == null || value.isEmpty) {
+                      return "This field is required.";
+                    } else if (value.length <= 10) {
+                      return "Please input the correct number.";
+                    } else if (value.contains(RegExp(r'[A-Z, a-z]'))) {
+                      return "Must contain only numbers.";
+                    }
+                    return null;
+                  },
+                ),
                 const SizedBox(height: 12),
-                createTextField(departmentController, 'Department', (value) {
-                  if (value == null || value.isEmpty) {
-                    return "This field is required.";
-                  } else if (value.contains(RegExp(r'[0-9]'))) {
-                    return "Must contain only letters";
-                  }
-                  return null;
-                }),
+                createTextField(
+                  departmentController,
+                  'Department',
+                  (value) {
+                    if (value == null || value.isEmpty) {
+                      return "This field is required.";
+                    } else if (value.contains(RegExp(r'[0-9]'))) {
+                      return "Must contain only letters";
+                    }
+                    return null;
+                  },
+                ),
                 const SizedBox(height: 16),
                 ElevatedButton(
                   onPressed: () {
@@ -177,20 +210,26 @@ class _SignUpState extends State<SignUp> {
                     }
                   },
                   style: ButtonStyle(
-                      minimumSize: MaterialStateProperty.all<Size>(
-                          const Size.fromHeight(60)),
-                      backgroundColor: MaterialStateProperty.all<Color>(
-                          const Color(0xFF081631)),
-                      shape: MaterialStateProperty.all<RoundedRectangleBorder>(
-                          RoundedRectangleBorder(
+                    minimumSize: MaterialStateProperty.all<Size>(
+                      const Size.fromHeight(60),
+                    ),
+                    backgroundColor: MaterialStateProperty.all<Color>(
+                      const Color(0xFF081631),
+                    ),
+                    shape: MaterialStateProperty.all<RoundedRectangleBorder>(
+                      RoundedRectangleBorder(
                         borderRadius: BorderRadius.circular(10.0),
-                      ))),
-                  child: const Text('SIGN UP',
-                      style: TextStyle(
-                        fontSize: 16.0,
-                        color: Colors.white,
-                        fontWeight: FontWeight.bold,
-                      )),
+                      ),
+                    ),
+                  ),
+                  child: const Text(
+                    'SIGN UP',
+                    style: TextStyle(
+                      fontSize: 16.0,
+                      color: Colors.white,
+                      fontWeight: FontWeight.bold,
+                    ),
+                  ),
                 ),
                 const SizedBox(height: 2),
                 Row(
@@ -198,24 +237,34 @@ class _SignUpState extends State<SignUp> {
                   children: [
                     const Text(
                       'Already have an account?',
-                      style:
-                          TextStyle(fontSize: 16.0, color: Color(0xFF777777)),
+                      style: TextStyle(
+                        fontSize: 16.0,
+                        color: Color(0xFF777777),
+                      ),
                     ),
                     TextButton(
-                        onPressed: () {
-                          Navigator.push(context,
-                              MaterialPageRoute(builder: (context) => LogIn()));
-                        },
-                        style: ButtonStyle(
-                          //this padding is the distance between ...account? and Sign in text
-                          padding: MaterialStateProperty.all<EdgeInsets>(
-                              const EdgeInsets.all(5.0)),
+                      onPressed: () {
+                        Navigator.push(
+                          context,
+                          MaterialPageRoute(builder: (context) => LogIn()),
+                        );
+                      },
+                      style: ButtonStyle(
+                        //this padding is the distance between ...account? and Sign in text
+                        padding: MaterialStateProperty.all<EdgeInsets>(
+                          const EdgeInsets.all(5.0),
                         ),
-                        child: const Text('Sign in',
-                            style: TextStyle(
-                                fontSize: 16.0, color: Color(0xFF081631))))
+                      ),
+                      child: const Text(
+                        'Sign in',
+                        style: TextStyle(
+                          fontSize: 16.0,
+                          color: Color(0xFF081631),
+                        ),
+                      ),
+                    )
                   ],
-                )
+                ),
               ],
             ),
           ),
@@ -223,40 +272,4 @@ class _SignUpState extends State<SignUp> {
       ),
     );
   }
-}
-
-/*
-*
-* Widget components
-*
-*
-* */
-Widget createTextField(valueController, label, validationFunction) {
-  return TextFormField(
-    validator: validationFunction,
-    controller: valueController,
-    decoration: InputDecoration(
-      labelText: label,
-      contentPadding: const EdgeInsets.symmetric(horizontal: 15),
-      labelStyle: const TextStyle(
-          color: Color(0xFFABABAB),
-          fontSize: 14), //affect the size of textfield
-      floatingLabelStyle: const TextStyle(color: Color(0xFF081631)),
-      //when textField is focused or selected
-      focusedBorder: const OutlineInputBorder(
-          borderRadius: BorderRadius.all(Radius.circular(10.0)),
-          borderSide: BorderSide(width: 2, color: Color(0xFF081631))),
-      //normal state of textField border
-      enabledBorder: const OutlineInputBorder(
-          borderRadius: BorderRadius.all(Radius.circular(10.0)),
-          borderSide: BorderSide(color: Color(0xFFABABAB))), // your color
-      //border style when error
-      errorBorder: const OutlineInputBorder(
-          borderRadius: BorderRadius.all(Radius.circular(10.0)),
-          borderSide: BorderSide(color: Color(0xFFFF0000))),
-      focusedErrorBorder: const OutlineInputBorder(
-          borderRadius: BorderRadius.all(Radius.circular(10.0)),
-          borderSide: BorderSide(width: 2, color: Color(0xFFFF0000))),
-    ),
-  );
 }
