+++ lib/core/dashboard.dart
@@ -3,8 +3,7 @@ import 'package:attendifyyy/api_connection/api_connection.dart';
 import 'package:attendifyyy/attendance/attendance_list.dart';
 import 'package:attendifyyy/authentication/user_preferences/user_preferences.dart';
 import 'package:attendifyyy/create_schedules/create_schedule.dart';
-import 'package:attendifyyy/create_students/create_students.dart';
-import 'package:attendifyyy/create_subjects/create_subjects.dart';
+import 'package:attendifyyy/subject_screen/list_of_subject.dart';
 import 'package:flutter/material.dart';
 import 'package:http/http.dart' as http;
 import 'package:intl/intl.dart';
